Refactor : Sliders 공용컴포넌트화 성공
