--- src/components/ProductCard/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ProductCard';
