+++ src/components/Sliders/Sliders.tsx
@@ -1,8 +1,8 @@
-import CategorySectionItem from '@/app/(provider)/(root)/_components/CategorySection/CategorySectionItem';
 import { Product } from '@/types/products';
 import { Tables } from '@/types/supabase';
-import ProductCard from '../ProductCard';
 import { Carousel, CarouselContent, CarouselItem } from '../ui/carousel';
+import CategoryCard from './CategoryCard';
+import ProductCard from './ProductCard';
 type SlidersProps = {
   data: Tables<'Brands'>[] | Product[];
   count: number;
@@ -23,9 +23,9 @@ const Sliders = ({ data, count }: SlidersProps) => {
           <CarouselItem className="flex-row-10" key={index}>
             {brandChunk.map((data) =>
               'productId' in data ? (
-                <ProductCard key={data.productId} product={data as Product} />
+                <ProductCard key={data.productId} product={data} />
               ) : (
-                <CategorySectionItem key={data.brandId} brand={data as Tables<'Brands'>} />
+                <CategoryCard key={data.brandId} brand={data} />
               )
             )}
           </CarouselItem>
